pyproject.toml
[project]
name = "certpy"
version = "0.1.0"
description = "Self-signed SSL certificate generator"
authors = [
    {name = "aprilahijriyan", email = "hijriyan23@gmail.com"},
]
dependencies = [
    "pyOpenSSL>=22.1.0",
    "typer>=0.6.1",
    "PyYAML>=6.0",
    "rich>=10.11.0",
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]

certpy = "certpy.__main__:cli"

[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "ipython>=7.16.3",
    "pre-commit>=2.9.2",
]
[tool.isort]
profile = "black"
